refactor: extract delegates from BrowserActivity and reorganize ConfigManager
Break up BrowserActivity (3220→2682 lines) by extracting three delegate classes:
- TranslationDelegate: 15 translation methods (translate, translateByParagraph, etc.)
- TabManager: 17 tab management methods (addAlbum, showAlbum, removeAlbum, etc.)
- FileHandlingDelegate: 10 file/EPUB/archive handling methods

Reorganize ConfigManager (1288→1135 lines) by extracting:
- PreferenceDelegates.kt: BooleanPreference, IntPreference, StringPreference,
  GestureTypePreference delegates and toggle() extension
- PreferenceEnums.kt: All preference-related enums (TranslationMode, FontType,
  DarkMode, HighlightStyle, etc.)
